\name{partycommonhtml}
\alias{partycommonhtml}
\title{ Function to produce HTML at the end of all EDM HTML pages}
\description{
	Often at the end of an HTML file code which is common across a set of pages is put.
	For example, a link back to some contents page. This function produces such code
	for the EDM example.
}
\usage{
partycommonhtml(file, append)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{ The HTML file to append the code to  }
  \item{append}{ Should always be TRUE }
}
\details{
	Generates HTML code which is going to be common to all pages generated by a given
	\code{\link{pinktoe}} run.
}
\value{ No value}
\references{ 
\url{http://www.stats.bris.ac.uk/~magpn/Research/Pinktoe/Welcome.html}
}
\author{Guy P Nason }
\note{ Not intended for general user use}

\seealso{ \code{\link{pinktoe}}}
\examples{
#
# A simple example where filename="" to show result on the screen
#
partycommonhtml("", TRUE)
# <a href="http://www.stats.bris.ac.uk/~magpn/Research/Politics/whovote.htm">Return to <EM>Who to Vote For</EM> page</a>
}
\keyword{ datagen }% at least one, from doc/KEYWORDS
