\name{writehtml}
\alias{writehtml}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Write the HTML file for an EDM decision and return the names of the called perl files so that they can be constructed. }
\description{
	Part of the pinktoe suite this function writes the HTML files that a web server will serve. The content of these HTML
    files will be some explanatory text (generated by edmtextfn, edmtittext), with a check box that will record the answer to
    the question specified in the HTML file. These HTML files will be able to call perl files and their names are constructed
    by this function and returned (and later used by writeperl). 
}
\usage{
writehtml(edmlabel, uniqedmlab, edmtextfn, edmtittext, cgibindir, localdir, commonhtml, numvar, numstring, treeobj)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{edmlabel}{ The variable name (from the \code{var} field of the original tree) for this HTML file/question. }
  \item{uniqedmlab}{ HTML file is associated with a node which is a given split on a variable: a unique filename for each
	node should be supplied (genedmhtml generates this by combining the variable name with the node number).} 
  \item{edmtextfn}{ A function that can print out explanatory text given a variable name (see web help page and \code{\end{displaymath}.text} example. }
  \item{edmtittext}{ Prints out a title string given a variable name }
  \item{cgibindir}{ A character string containing the directory where the perl
        files generated by pinktoe (with the extension \code{.pl}) will be
        stored.
        (This should be the directory part of the URL of the cgi-bin
        directory). }

  \item{localdir}{A local location to store both the HTML and perl files
        immediately after they are generated }

  \item{commonhtml}{A user-supplied function that prints out some HTML code.
        This is appended to every HTML web page.}

  \item{numvar}{ Logical variable which is TRUE is the variable for this HTML is numeric, FALSE if it is a factor. }
  \item{numstring}{ The S character string which describes the split (either factor or numeric) for this variable. Basically one of the cells in the
		dataframe returned by \link{splits.rpart}}
  \item{treeobj}{ The rpart (or tree) object that this whole process is concerned with}
}
\details{
	Writes the HTML files
}
\value{The HTML file here creates a checkbox. The result of which causes a call to a perl function. The HTML file here creates the perl file name
	and passes it back to its calling function (\code{\link{genedmhtml}}) which then creates the perl function and supplies it with the file name.
	This function returns the perl filename.
}
\references{ 
\url{http://www.stats.bris.ac.uk/~magpn/Research/Pinktoe/Welcome.html}
}
\author{Guy P Nason }
\note{ Not intended for general user use}

\seealso{ \code{\link{pinktoe}}}
\examples{
#
#Not intended for general user use
#
}
\keyword{ internal }% at least one, from doc/KEYWORDS
